// src/types/app.ts
export interface MovieListProps {
  title: string;
  path: string;
  coverType: "poster" | "backdrop";
}

export interface Movie {
  backdrop_path: string;
  genres: { id: number; name: string };
  homepage: string;
  id: number;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: number;
  original_language: string;
  production_companies: {
    id: number;
    logo_path: string;
    name: string;
    origin_country: string;
  };
  production_countries: {
    iso_3166_1: string;
    name: string;
  };
  release_date: Date;
  revenue: number;
  runtime: number;
  spoken_languages: {
    english_name: string;
    iso_639_1: string;
    name: string;
  };
  status: string;
  tagline: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface MovieItemProps {
  movie: Movie | IMovie;
  size: { width: number; height: number };
  coverType: "poster" | "backdrop";
}

export interface IMovie {
  adult: boolean;
  backdrop_path: string;
  belongs_to_collection: BelongsToCollection;
  budget: number;
  genres: Genre[];
  homepage: string;
  id: number;
  imdb_id: string;
  origin_country: string[];
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  production_companies: ProductionCompany[];
  production_countries: ProductionCountry[];
  release_date: string;
  revenue: number;
  runtime: number;
  spoken_languages: SpokenLanguage[];
  status: string;
  tagline: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface BelongsToCollection {
  backdrop_path: string;
  id: number;
  name: string;
  poster_path: string;
}

export interface Genre {
  id: number;
  name: string;
}

export interface ProductionCompany {
  id: number;
  logo_path: string;
  name: string;
  origin_country: string;
}

export interface ProductionCountry {
  iso_3166_1: string;
  name: string;
}

export interface SpokenLanguage {
  english_name: string;
  iso_639_1: string;
  name: string;
}

export interface IMovieDetail
  extends Pick<
    IMovie,
    | "adult"
    | "backdrop_path"
    | "id"
    | "original_language"
    | "original_title"
    | "overview"
    | "popularity"
    | "poster_path"
    | "release_date"
    | "title"
    | "video"
    | "vote_average"
    | "vote_count"
  > {
  genre_ids: number[];
  media_type: string;
}

export interface IMovieSearch extends Omit<IMovieDetail, "media_type"> {}